Add files via upload
This student management system was developed using advanced Java with Eclipse as the integrated development environment and MySQL Workbench for database management. The project involves creating a database to store student information and utilizes Java programming to implement features such as inserting, deleting ,updating student name, phone number and many more.

The primary learning aspects of this project include database creation in MySQL Workbench, coding and application development in Eclipse, and understanding the interaction between the Java application and the MySQL database.